requirements.txt
anthropic>=0.86.0
pandas
pydantic
python-dotenv
